pyproject.toml
[project]
name = "study_coach_agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.21.0",
    "litellm>=1.80.11",
]
